backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic>=2.0
pydantic-settings>=2.0
supabase>=2.0
httpx>=0.27
celery>=5.3
redis>=5.0
groq>=0.9
scikit-learn>=1.4
pandas>=2.2
numpy>=1.26
joblib>=1.4
shap>=0.45
cryptography>=42.0
python-jose[cryptography]>=3.3
slowapi>=0.1.9
sentry-sdk[fastapi]>=2.0
structlog>=24.0
pytest>=8.0
pytest-asyncio>=0.23
